add first prototype of httpErrorfilter
