+++ .gitignore
@@ -1,3 +1,3 @@
 input/
-__pychache__/
 output/
+FilterManager/__pychache__/
